Replace "once" and "later" with new syntax features "\=" and "\==".
This allows you to do an "inline" once-value, instead of having to evaluate it
first to create a handle to it.  For example, let's say you have a "show"
function which says its argument three times:

	\show=(\x say x say x say x)

Formerly, to create a once-value, you would first have to create a named handle
to it by evaluating it with the named "once" function:

	\x=(once; put "(CALC)" + 2 3)
	show x

That prints:

	(CALC)5
	5
	5

With the new "\=" syntax, you can create the once-value on the fly:

	show (\= put "(CALC)" + 2 3)

This feature can make your code far more concise and clear, because you don't
have to introduce a name every time you want a once-value.

Also, the "\==" syntax is a replacement for the named "later" function, which
does deferred evaluation.

Note that I have not yet eliminated the use of "once" and "later" in the test
code, because I wanted this version to have minimal differences from 15.0.0.
I will do that in the next version, and you will see how much more concise the
code looks.

As it turns out, the now-deprecated "once" and "later" functions can be defined
using the new syntax features as:

	\once=(\x \== \= x)
	\later=(\x \== x)

To demonstrate this, I have actually redefined them that way in main.fxl.  I
could have gone ahead and eliminated them as built-in C functions, but that
would break "fexl0" which bypasses main.fxl.  In a future version 16.0.0 I will
remove the names entirely.