Eliminate deprecated features.
I eliminated the "\==" syntax for "later" or deferred evaluation.  In fact, I
eliminated the "later" concept altogether.  If you want a function to return
another function that should only be evaluated later, you should "yield" it to
the caller (i.e. through a continuation).

For example:

	\get_function=(\key\val yield; say [key" = "val])

	get_function "x" 23 \say_x
	get_function "y" 42 \say_y

	say_x
	say_y

The output is:

	x = 23
	y = 42

The yield function is defined in the C code, but can be defined in Fexl as:

	\yield=(\x\f f x)

The "define" function, which defines a symbol in a context, formerly used a
"later" value:

	\define=(\key\val with key (later val))

Now instead it uses {val}, which is equivalent to (yield val) but marginally
faster:

	\define=(\key\val with key {val})

I eliminated the deprecated function g_sort.  The sort_items function is far
more efficient in the general case.

I eliminated the deprecated function stream_list, which was a synonym for
get_list.

I eliminated the deprecated "crypto_" synonyms, which were mapped as follows:

	crypto_box_public  => nacl_box_public
	crypto_box_prepare => nacl_box_prepare
	crypto_box_seal    => nacl_box_seal
	crypto_box_open    => nacl_box_open
	crypto_sign_public => nacl_sign_public
	crypto_sign_seal   => nacl_sign_seal
	crypto_sign_open   => nacl_sign_open
	crypto_hash        => sha512