Chained modules
This is version 17.0.0, which is not compatible with the version 16 series.

This version allows you to chain modules together using the ":" function,
placing them in the exact search order you desire.  The module is only loaded
if a symbol is not found in any previous module, and is cached so subsequent
uses of the same module do not read the source file again.