Implement "with" and "fetch" directly in C.
This improves the performance of programs which make heavy use of caching.  The
performance was fine before, but doing those two crucial routines in C is an
easy win.  In one strenuous test I saw a 30% improvement in speed and a 36%
improvement in memory usage.

I also added a "cache_1" function which is equivalent to (remember 1).