Optimized list representation
I added a type_list combinator which allows a compressed representation of
lists.  Formerly, the list form [x1;xs] was represented in memory directly in
terms of the cons function, with this structure:

	[type_cons [type_cons Qcons x1] xs]

The new equivalent structure is:

	[type_list Qlist [0 x1 xs]]

When you count the left-brackets to see how many values are used, it is 2 in
both cases, so it doesn't appear to save any memory.  However, the benefits
accrue when the list form starts with more items.

Previously, the list form [x1 x2;xs] had this structure:

	[type_cons [type_cons Qcons x1]
	[type_cons [type_cons Qcons x2] xs]]

The new structure is:

	[type_list Qlist [0 x1 [0 x2 xs]]]

The old structure uses 4 values, but the new structure only uses 3.

For larger lists, the benefits are higher.  For example, the list form
[x1 x2 x3 x4 x5;xs] had this structure:

	[type_cons [type_cons Qcons x1]
	[type_cons [type_cons Qcons x2]
	[type_cons [type_cons Qcons x3]
	[type_cons [type_cons Qcons x4]
	[type_cons [type_cons Qcons x5] xs]]]]]

The new structure is:

	[type_list Qlist [0 x1 [0 x2 [0 x3 [0 x4 [0 x5 xs]]]]]]

Here the old form uses 10 values, but the new form only uses 6.

In general, for a list form starting with N items, the number of values used in
the old and new representations are:

	old:  2*N
	new:  N+1

For large lists used in data files, the savings are significant, considering
that each value is 32 bytes.  For a list with 10000 items, you save 319968
bytes.

The new representation also makes the functions in type_tuple.c more efficient
and compact.  Note how that source file has dropped from 82 lines to 49.