Simplify representation of forms and symbols.
The operations in type_sym.c formerly used specialized structures with inline
modification to track the free symbol lists in forms.  Now I do it in a more
functional style, without inline modification.

This makes it easier to verify the code, but also prepares the way for taking a
new approach to contexts.  In the new approach, a context will not be a map
from symbol to its definition.  Instead, a context will be a function that
defines 0 or more symbols in an unresolved form.  In other words, I will allow
partial resolution of forms.  I believe this will greatly simplify the use of
contexts, no longer requiring "chaining" and allowing fixed, cache-able context
functions.