Clean up symbol and form structs.
Now I store the symbol name as a value instead of directly as a string.  This
simplifies resolve_name, and is also needed for the new way of handling contexts
coming in version 23.

I also put the source label on each symbol instead of at the form level.  This
cleans up some code, and is also more in line with the way of doing things in
version 23.