Implement "lib" function which loads libraries only once on demand.
The "using" function evaluates a form with a list of contexts chained together.
This eliminates the need for "::".  The "chain" function is available directly
if you need it.

The "def_lib" function creates a library system, and is used for both the
system libraries in dir_lib and the user's libraries in dir_local.  The b48.fxl
test demonstrates using def_lib with a simulated file system.