Eliminate "def_lib".
I think using a cache to load libraries once on demand is overrated, and the
semantics are too complicated.  If you want to load a library once, you can do
it easily by thinking ahead a little.

I'm inclined to go even further in the future and eliminate all uses of var,
even for memoized functions.  You can do this instead by passing around a table
of precomputed results.

Eliminated:

	chain
	using
	def_lib
	def_lib_in