Restore numeric constants predefined in the parser.
This handles numeric constants the same way as in version 22, predefined in the
parser instead of resolved by a context function.  Doing that with "def_num" was
slow and overly complicated.

I also changed struct symbol to contain a value for the name instead of a
direct string.  That simplifies resolve_name and also fixes a segmentation
fault that could occur if you defined a context that returned a name itself as
the value.  I included a regression test for this in test/b13.fxl.