Define context in terms of partial resolution of forms.
Formerly, a context was defined as:

	a function that maps a symbol to its definition

Now a context is defined as:

	a function that defines zero or more symbols in a form

This simplifies and accelerates the resolution of symbols, providing greater
flexibility and economy of expression.

I have done lower-level performance measurements which demonstrate a large
percentage increase in efficiency.  The old way was not particularly slow, but
it's good to see that the new way is both simpler and faster.

New definitions added:

	def
	value
	resolve
	std
	is_closed
	use_file
	use
	defc
	::

Old definitions dropped:

	def
	evaluate
	resolve
	std
	use_file
	parse_file
	use
	def_context
	::
	import
	import_as