Normalize the order of built in libraries in main.fxl.
Now the libraries are listed in the order of their definition, eliminating the
need to think about it.  That happened to save 7K of memory on test/b23.fxl.

I also added a test/load.fxl which demonstrates a technique for loading modules
once on demand by putting the dependency graph in one place.  This could help
with managing a large project with many interrelated files.