Restore the fexl0 feature.
The comment on the eval_script routine in type_standard.c now reads as follows:

~~
Evaluate the user's script.  Read the script from the file named by argv[1] if
present, or from stdin otherwise.

If this program is running as "fexl0", it runs your script directly.  If it is
running as "fexl", it first runs the local script "src/main.fxl", which then
runs your script within an extended library context.

The purpose of "fexl0" is to give you the option of bypassing src/main.fxl
altogether, defining any extensions beyond the built-in C standard context as
you please.

In a future release I will make fexl0 the default behavior, so it always reads
your script directly without loading any default library.  Predefined libraries
such as the current main.fxl will be provided in the lib directory so you can
pull them in with a one-liner.

I am making this change because the one-size-fits-all approach to a standard
library is a losing proposition.  Eventually I might even factor out the
currently built-in C functions into independent libraries loaded on demand.
Then you could build an entirely separate C library, put it in your own
directory somewhere, and load it from within Fexl.  That would even give you
the power to generate binary code on the fly and run it immediately.
~~