Use '\#' instead of '\\' to simulate end of file.
I'm doing this because I'm moving toward a cleaner syntax for simple and eager
definitions, and I need to use '\\' for a better purpose.  Besides, '\#' makes
more sense because it resembles the '#' comment feature.

The following examples illustrate the current syntax for definitions:

	\x==(+ 2 5)   # simple definition
	f x x

	\x=(+ 2 5)    # eager definition
	f x x

	\x
	\x=x          # force evaluate parameter
	f x x

The following examples illustrate the future syntax:

	\x=(+ 2 5)    # simple definition
	f x x

	\\x=(+ 2 5)   # eager definition
	f x x

	\\x           # force evaluate parameter
	f x x

As you can see, the future syntax uses '\\' to represent eager definition.
This scheme will enable me to eliminate the '==' syntax altogether, finally
standardizing on a single '=' for all definitions, simple or eager.

I am also deprecating "eval" as a function that can be explicitly called.
There is no need to call it directly because we already have a syntax for eager
definitions.

In the current syntax, eager definitions make use of type_eval in basic.c:

	(eval x f) = (f y), where y is the final value of x.

In the future syntax, eager definitions will make use of a new type_bang, which
is just like eval except the parameters are swapped:

	(bang f x) = (f y), where y is the final value of x.

The swapping of the parameters is crucial for implementing the '\\x' lambda
form with no '=' sign, which forces evaluation of the incoming parameter.