Keeping 26.13.0.
The changes proposed in version 27 might not be a good idea.  In the current
version 26, a "\x=D" definition is eager, meaning D is evaluated first.  That's
usually what I want, for example:

	\fh=(fopen "file" "r")

	\x=(+ y (* z z))

In the first case, I definitely want to open the file right now.  In the second
case, I might as well go ahead and evaluate the expression because it's likely
that x is used more than once somewhere below.

If '=' were to indicate a lazy definition, that could cause unhappy accidents.
Having to remember to use the proposed '\\' for eager would be too much mental
baggage.  The '=' is also consistent with C semantics, making it a natural fit.

In the current version 26, it's easy to make a lazy definition when I want it
just by changing the '=' to '=='.

I had also proposed using '\\' to force evaluation of an incoming argument, as
in:

	\\x
	f x x

In the current version 26, I have to say:

	\x
	\x=x
	f x x

I think that's just the way it has to be.  Besides, in cases where you evaluate
an argument conditionally, you'd need to use that technique anwyay.