Added functions for object introspection.
An "object" is a function which maps a key (string or number) to an arbitrary
value.  Like any function, it can be defined as a lambda form, but an object is
often defined using the "with" function to map keys to values.  An object of
this kind is known as an "assoc" object, and it operates more efficiently than
a lambda form.  Its internal type is type_assoc.

I have now added these new functions:

	is_obj
	split_obj

The is_obj function determines if a value is an assoc object (type_assoc).

The split_obj function splits an assoc object into its components key, val, and
tail, where key and val are the first pair in the object, and tail is the value
following the first pair.  Typically the tail is either another object or the
final value void.

I added some tests to test/b45.fxl to illustrate how these are used.  There I
defined a val_str function which converts a value to a string.  The value can
be a scalar, a list or tuple of values, or an assoc object.

I did this so I could cache object values in files.  The val_str function
converts any object to a string which I can write into a file.  Without it, I
would need to write specialized conversion routines for every type of object I
used, requiring me to keep track of the exact list of keys in each type.