Add functions to read dynamic libraries.
Functions added:

	dlopen
	dlsym

I added "test/lib.fxl" as an example, which dynamically loads the "show"
function from a new library "lib/show.so" and calls it on several test cases.

It is now possible for a Fexl program to create and run C code on the fly.
To do this it would write the code into a file, compile it into a shared
library, load the library with dlopen, grab the new functions with dlsym,
and run them.