Minor improvements
I now call skip in parse.c to set the initial ch.  Formerly I was setting it to
' ' (space).  This brings it in line with some other code I'm developing to
accelerate parsing in Fexl.  Setting the inital ch to ' ' works for the Fexl
syntax, but it does not work in general for all parsing tasks.

I am now using is_undef instead of is_void in some libraries and tests.  It's
slightly quicker to use is_undef when you know the argument is already
evaluated.