Use procedural approach in parsers
The lib/read_csv.fxl and lib/read_ssv.fxl modules now use a procedual approach
to parsing.  This is more conventional than the purely functional approach of
reading a stream of characters, and is more in line with how the Fexl parser
itself works for example.

This uses a new lib/read.fxl module which provides two functions "look" and
"skip".  The look function returns the current character in the input stream.
The skip function advances to the next character.  It currently uses vars to
maintain this state.  I've put a note in there saying that I intend to code
this in C to make it faster, but it's quite acceptable as is.

The updated parsers also avoid the technique of passing in continuation
functions such as "no" and "yes".  Instead, they just return ordinary values
which in some cases can be void or boolean.

I think this kind of code is far easier to understand and manipulate.  The old
purely functional approach used a "point-free" or "monadic" technique to hide
the stream being passed around, and that can be difficult to comprehend.
There's a lot to be said for plain procedural programming, and Fexl has
always been amenable to that.