Avoid incrementing num_steps when not benchmarking.
Formerly the eval routine in value.c would increment num_steps on every
reduction step, regardless of whether it was running under fexl_benchmark.
Now it does not.  Instead, fexl_benchmark swaps out the normal eval routine
with a new version that counts the steps, and then restores it when done.

To do that, I had to make eval into a function pointer so it could be swapped
dynamically.  I was concerned about function pointer overhead compared with
incrementing num_steps every time, so I did some timing experiments with a
script that looped 16,000,000 times.

With no benchmarking in play, the times averaged:

	OLD : 3.756 s
	NEW : 3.625 s

So the new version was 3.6% faster when running normally.

With benchmarking, the times averaged:

	OLD : 3.819 s
	NEW : 3.942 s

So the new version was 3.1% slower when running under fexl_benchmark.  That's
fine because I wanted to accelerate normal evaluation.

Note that if I disabled fexl_benchmark entirely then eval would not have to be
a function pointer, and then the ideal run time would be about 3.534 s.

The old version is 6.3% slower than the ideal time, which reflects the overhead
of incrementing num_steps every time.

The new version is 2.6% slower than the ideal time, which reflects the overhead
of calling eval through a function pointer.

Clearly the new version narrows the gap with respect to the ideal time, and I
am willing to pay the 2.6% to keep fexl_benchmark because it's a simple and
portable way to measure reduction steps and memory usage during Fexl evaluation.