Some refactoring in parse.c
1. I changed the at_white routine slightly.

The old code defines it as:

	return isspace(cur_ch) || iscntrl(cur_ch);

I included iscntrl because I wanted NUL and other control characters to be
considered white space.  Interestingly, that includes the 127 (DEL) character.

The new code defines it as:

	return cur_ch <= ' ' && cur_ch != -1;

Note that -1 is the only possible negative value.

So here is how white space is defined:

	old code : 0-32, 127
	new code : 0-32

The net effect is to exclude DEL as a white space character, which allows
parse_name and parse_terminator to use a simple (cur_ch > ' ') test.

2. I simplified the structure of parse_exp.