Fix bug in parse_tuple.
When parsing the empty tuple "{}", it was not wrapping it with Qtuple.  That
made {} equivalent to [].  It should be equivalent to I, and now it is.

When parsing the empty list "[]", that always worked fine because there is no
need to wrap [] with Qlist.

I added a test case to test/a1.fxl.