Eliminate expansion of lists in place.
This eliminates the "expand" routine, which expands a list in place.  That was
a misguided concept.  Instead, I now use an "iterate" routine to run through
evaluable lists without affecting the original values.

I use iterate in type_tuple, which applies an expression to the list of values
in a tuple.  I also use it in beg_collect, which is called by type_exec to
gather the vector of string values passed to the execv call.