Clarify naming of pipes in op_process.
A pipe p is a pair of end points, reading from p[0] and writing to p[1].

The name of an end point is two letters indicating who and where:

	(c:child p:parent) x (i:in o:out e:err)

The name of a pipe is the read and write end points joined by an underscore.
The op_process routine uses:

	ci_po : read child in, write parent out
	pi_co : read parent in, write child out
	pe_ce : read parent err, write child err

An index 0 or 1 denotes the read or write end point of a pipe, for example:

	ci_po[0] : child in
	ci_po[1] : parent out

These naming conventions greatly clarify the code.