Properly handle forms with table->count == 0.
I now use a common routine "is_closed" in type_sym.c to check if a table is
closed (fully resolved).  Formerly I assumed that a closed table would always
be a null pointer, but that is not true even in the simple example (\; \x x).
In that case the table is not null and table->count == 0.

This fixes bugs in type_closed and type_value, and I added a case in
test/a1.fxl to check this.

The op_resolve routine also calls is_closed, which is a slight optimization of
the table->count == 0 case but was not a bug previously.