Expand list data inline.
I reverted to an earlier technique of expanding list data inline, bringing back
the expand routine into basic.c.  I then deleted the iterate routine, which
iterated over a list without inline expansion.

As it turns out, iterate was not completely reliable, particularly with fancy
lists involving embedded "once" constructs and such.  I added two test suites
to test/a1.fxl which would cause a segmentation fault in the previous version
26.30.0+1.  They now work correctly in this new version 26.30.1.

I also added test.so to test a call to op_argv that actually returns, unlike
type_exec which never returns.

The op_argv routine no longer dies with an error message if it sees an item
that is not a string.  It now simply ends the list at that point.