Restore inline record updates.
In the previous version 26.31.1 I said that doing inline record updates was a
"bad idea," because I found a way to generate a circular structure, yielding a
semantically invalid result and a memory leak.

As it turns out, it's a great idea, not a bad one, and the problem is easily
solved.  Now when you call (set key val obj), the val is evaluated before being
stored in the record.  It was the lazy val that allowed the problem to occur.

Because that behavior is not strictly equivalent to version 26, I have bumped
the major version to 27 here.  Happy New Year!