Fix bug in test/b15.fxl
The index_del function in test/b15.fxl had a bug where it could drop an entire
branch prematurely.  That is fixed now, and I added test_5 to check it.

In another branch I'm working on a drastic simplification of the index code
anyway, but I wanted to go ahead and fix it here.