Add line number to reference expressions
I'm not using this yet, but I'm doing it in preparation for deferred symbol
resolution.  That will avoid looking up symbols in the current context during
the parse.  Instead, it will return an expression with the unbound symbols
still in the map, and those will be resolved outside the parse.

This will enable me to implement the "\;" syntax for forms currently in the
master branch.