Eliminate inline list expansion in type_tuple.
I never really liked the concept of expanding a list inline using the expand()
routine, but it seemed to work.  As it turns out, it can cause a weird memory
error in some obscure cases.  So it really is a bad concept.

It is semantically cleaner and more reliable to march through the list without
changing the original.  I now do this in type_tuple, and I have added some
code in test/a1.fxl to test the obscure cases.

I still use the expand() routine in one place in type_str.c, so I have moved it
there in preparation for eliminating it.