A context is a function from string to value.
Formerly a context was a function that defined some symbols in a form,
returning a new form.

Now a context is simply a function which maps a string name to a value.  For
a given name, it returns (yield val) if the name is defined, or (void) if not
defined.

I have added these new functions:

(cx_std)
A function which maps a name in the standard context to its value, if any.

(define key val cx)
Define key as val in context cx.  Equivalent to:

	\define=(\key\\val with key (yield val))

(resolve cx form)
Use context cx to resolve some symbols in the form, returning a new form.

(evaluate cx form)
Resolve all the symbols in the form and return its value.  If there are any
undefined symbols, report them all and die.

(extend cx form)
Same as evaluate except it also defines the name "cx_std" to refer to the
context cx.  Equivalent to:

	\extend=(\cx evaluate (define "cx_std" cx; cx))

NOTE:  This version of the code is a "bridge" version, preserving backward
compatibility.  The following symbols are now deprecated:

	std
	def
	value
	::

I will eliminate the deprecated symbols from my existing code, using these
instead:

	cx_std
	define
	evaluate
	chain

After that is done, I will likely rename as follows:

	cx_std   => std
	define   => def
	evaluate => value
	chain    => ::

During the port of existing code to the new context concept, there is a neat
trick you can use to convert a new-style context back to an old-style context:

	\old_cx=(resolve new_cx)

For example a new-style context might be:

	\new_cx=
		(
		define "fred" (say "I am Fred.");
		define "wilma" (say "I am Wilma.");
		void
		)

The old-style context is:

	\old_cx=
		(\form
		def "fred" (say "I am Fred.");
		def "wilma" (say "I am Wilma.");
		form
		)

You can easily eliminate the old-style context with this transformation:

	\old_cx=
		(
		resolve;
		define "fred" (say "I am Fred.");
		define "wilma" (say "I am Wilma.");
		void
		)

Then you can percolate the resolve calls up out of that code.  At some point
you might go from something like this:

	(\form
	old_cx_A;
	old_cx_B;
	form
	)

To this:

	(\form
	resolve new_cx_A;
	resolve new_cx_B;
	form
	)

At that point you can apply this transformation:

	(
	resolve;
	chain new_cx_B new_cx_A
	)

In general you can go from this:

	(\form
	resolve new_cx_A;
	resolve new_cx_B;
	resolve new_cx_C;
	resolve new_cx_D;
	form
	)

To this:

	(
	resolve;
	chain new_cx_D;
	chain new_cx_C;
	chain new_cx_B;
	chain new_cx_A;
	void
	)

Note also you can always change (chain cx void) to simply cx:

	(
	resolve;
	chain new_cx_D;
	chain new_cx_C;
	chain new_cx_B;
	new_cx_A
	)

In the next version of the code I will port the test suite to the new context
concept.  For now I have left it intact to demonstrate backward compatibility.