Refactor stream code so that the save and restore is more obviously correct.
Now the code calls beg_stream to point to a new stream source, and that returns
the previous state of the stream.  Then after reading the stream it calls
end_stream to restore the previous state.  This ensures that stream processing
is always re-entrant.

The code was already doing the save and restore correctly, but now it is more
obvious.